Added object id validation on mongo repositories to avoid unhandled rejections

+++ src/modules/games/infra/mongoose/repositories/AchievementsRepository.ts
@@ -1,3 +1,7 @@
+import { isValidObjectId } from 'mongoose';
+
+import { RequestError } from '@shared/errors/implementations';
+import errorCodes from '@config/errorCodes';
 import { IAchievementsRepository } from '@modules/games/repositories';
 import { IAchievement } from '@modules/games/entities';
 import Achievement, {
@@ -16,19 +20,35 @@ export default class AchievementsRepository
     id: string,
     gameId: string,
   ): Promise<IAchievementDocument | undefined> {
+    if (!isValidObjectId(id))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     return await Achievement.findOne({
       _id: id,
       game: gameId,
     }).populate('title');
   }
 
-  public async create(
-    achievement: Omit<IAchievement, 'id'>,
-  ): Promise<IAchievementDocument> {
-    return await Achievement.create(achievement);
+  public async create({
+    name,
+    description,
+    game,
+    image,
+    title,
+  }: Omit<IAchievement, 'id'>): Promise<IAchievementDocument> {
+    return await Achievement.create({
+      name,
+      description,
+      game,
+      image,
+      title,
+    });
   }
 
   public async delete(achievementId: string, gameId: string): Promise<void> {
+    if (!isValidObjectId(achievementId))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     await Achievement.deleteOne({ _id: achievementId, game: gameId });
   }
 
@@ -39,6 +59,9 @@ export default class AchievementsRepository
     image,
     title,
   }: IAchievement): Promise<IAchievement> {
+    if (!isValidObjectId(id))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     const updatedAchievement = await Achievement.findByIdAndUpdate(
       id,
       {
