+++ src/modules/games/infra/mongoose/repositories/ActivitiesRepository.ts
@@ -1,3 +1,7 @@
+import { isValidObjectId } from 'mongoose';
+
+import errorCodes from '@config/errorCodes';
+import { RequestError } from '@shared/errors/implementations';
 import { IActivitiesRepository } from '@modules/games/repositories';
 import { IActivity } from '@modules/games/entities';
 import Activity, {
@@ -14,19 +18,35 @@ export default class ActivitiesRepository
     id: string,
     gameId: string,
   ): Promise<IActivityDocument | undefined> {
+    if (!isValidObjectId(id))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     return await Activity.findOne({
       _id: id,
       game: gameId,
     });
   }
 
-  public async create(
-    activity: Omit<IActivity, 'id'>,
-  ): Promise<IActivityDocument> {
-    return await Activity.create(activity);
+  public async create({
+    name,
+    description,
+    experience,
+    dmRules,
+    game,
+  }: Omit<IActivity, 'id'>): Promise<IActivityDocument> {
+    return await Activity.create({
+      name,
+      description,
+      experience,
+      dmRules,
+      game,
+    });
   }
 
   public async delete(activityId: string, gameId: string): Promise<void> {
+    if (!isValidObjectId(activityId))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     await Activity.deleteOne({ _id: activityId, game: gameId });
   }
 
@@ -39,6 +59,9 @@ export default class ActivitiesRepository
     experience,
     history,
   }: IActivity): Promise<IActivity> {
+    if (!isValidObjectId(id))
+      throw new RequestError('Id is invalid!', errorCodes.INVALID_ID);
+
     const updatedActivity = await Activity.findByIdAndUpdate(
       id,
       {
